changed file structure
